requirements.txt

streamlit
# mediapipe  # commented out: heavy build on Windows/Py3.14; enable after installing proper wheels or build tools
opencv-python
numpy
Pillow
